--- a/packages/angular/build/src/tools/angular/angular-host.ts
+++ b/packages/angular/build/src/tools/angular/angular-host.ts
@@ -164,6 +164,7 @@ export function createAngularCompilerHost(
   typescript: typeof ts,
   compilerOptions: AngularCompilerOptions,
   hostOptions: AngularHostOptions,
+  packageJsonCache: ts.PackageJsonInfoCache | undefined,
 ): AngularCompilerHost {
   // Create TypeScript compiler host
   const host: AngularCompilerHost = typescript.createIncrementalCompilerHost(compilerOptions);
@@ -229,16 +230,17 @@ export function createAngularCompilerHost(
     return hostOptions.modifiedFiles;
   };
 
+  // Provide a resolution cache to ensure package.json lookups are cached
+  const resolutionCache = typescript.createModuleResolutionCache(
+    host.getCurrentDirectory(),
+    host.getCanonicalFileName.bind(host),
+    compilerOptions,
+    packageJsonCache,
+  );
+  host.getModuleResolutionCache = () => resolutionCache;
+
   // Augment TypeScript Host for file replacements option
   if (hostOptions.fileReplacements) {
-    // Provide a resolution cache since overriding resolution prevents automatic creation
-    const resolutionCache = typescript.createModuleResolutionCache(
-      host.getCurrentDirectory(),
-      host.getCanonicalFileName.bind(host),
-      compilerOptions,
-    );
-    host.getModuleResolutionCache = () => resolutionCache;
-
     augmentHostWithReplacements(typescript, host, hostOptions.fileReplacements, resolutionCache);
   }